Allow building for web consecutively instead of parallely

+++ build-web/build.sh
@@ -4,9 +4,15 @@ set -e
 
 # Config
 
-# To speed up builds with single-threaded full LTO linking,
-# we run all builds in parallel each from their own folder.
-export NUM_JOBS=5
+source /root/common/prep.sh
+
+if [ "${BUILD_WEB_PARALLEL}" == "1" ]; then
+  # To speed up builds with single-threaded full LTO linking,
+  # we run all builds in parallel each from their own folder.
+  export NUM_JOBS=5
+else
+  export NUM_JOBS=1
+fi
 declare -a JOBS=(
   "target=editor use_closure_compiler=yes"
   "target=template_debug"
@@ -22,40 +28,54 @@ export TERM=xterm
 
 source /root/emsdk/emsdk_env.sh
 
-if [ ! -z "${PRESET_GODOT_DIR}" ]; then
-  cd $PRESET_GODOT_DIR
-  rm -rf bin
-else
-  rm -rf godot
-  mkdir godot
-  cd godot
-  tar xf /root/godot.tar.gz --strip-components=1
-fi
-
 # Classical
 
 if [ "${CLASSICAL}" == "1" ]; then
   echo "Starting classical build for Web..."
 
   for i in {0..4}; do
-    cp -r /root/godot /root/godot$i
-    cd /root/godot$i
-    echo "$SCONS platform=web ${OPTIONS} ${JOBS[$i]}"
-    $SCONS platform=web ${OPTIONS} ${JOBS[$i]} &
-    pids[$i]=$!
+    if [[ "${JOBS[$i]}" == *"editor"* ]]; then
+      if [ "${BUILD_EDITOR}" != "1" ]; then
+        continue
+      fi
+    else
+      if [ "${BUILD_TEMPLATES}" != "1" ]; then
+        continue
+      fi
+    fi
+
+    if [ "${BUILD_WEB_PARALLEL}" == "1" ]; then
+      cp -r /root/godot /root/godot$i
+      cd /root/godot$i
+      echo "$SCONS platform=web ${OPTIONS} ${JOBS[$i]}"
+      $SCONS platform=web ${OPTIONS} ${JOBS[$i]} &
+      pids[$i]=$!
+    else
+      cd /root/godot
+      echo "$SCONS platform=web ${OPTIONS} ${JOBS[$i]}"
+      $SCONS platform=web ${OPTIONS} ${JOBS[$i]}
+    fi
   done
 
-  for pid in ${pids[*]}; do
-    wait $pid
-  done
+  if [ "${BUILD_WEB_PARALLEL}" == "1" ]; then
+    for pid in ${pids[*]}; do
+      wait $pid
+    done
+  fi
 
-  mkdir -p /root/out/tools
-  cp -rvp /root/godot0/bin/*.editor*.zip /root/out/tools
+  if [ "${BUILD_EDITOR}" == "1" ]; then
+    mkdir -p /root/out/tools
+    [[ ${BUILD_WEB_PARALLEL} = 1 ]] && godot_dir="godot0" || godot_dir="godot"
+    cp -rvp /root/$godot_dir/bin/*.editor*.zip /root/out/tools
+  fi
 
-  mkdir -p /root/out/templates
-  for i in {1..4}; do
-    cp -rvp /root/godot$i/bin/*.zip /root/out/templates
-  done
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    mkdir -p /root/out/templates
+    for i in {1..4}; do
+      [[ ${BUILD_WEB_PARALLEL} = 1 ]] && godot_dir="godot${i}" || godot_dir="godot"
+      cp -rvp /root/$godot_dir/bin/*.zip /root/out/templates
+    done
+  fi
 fi
 
 # Mono
@@ -67,12 +87,14 @@ if false; then
 
   cp -r /root/mono-glue/GodotSharp/GodotSharp/Generated modules/mono/glue/GodotSharp/GodotSharp/
 
-  $SCONS platform=web ${OPTIONS} ${OPTIONS_MONO} target=template_debug
-  $SCONS platform=web ${OPTIONS} ${OPTIONS_MONO} target=template_release
+  if [ "${BUILD_TEMPLATES}" != "1" ]; then
+    $SCONS platform=web ${OPTIONS} ${OPTIONS_MONO} target=template_debug
+    $SCONS platform=web ${OPTIONS} ${OPTIONS_MONO} target=template_release
 
-  mkdir -p /root/out/templates-mono
-  cp -rvp bin/*.zip /root/out/templates-mono
-  rm -f bin/*.zip
+    mkdir -p /root/out/templates-mono
+    cp -rvp bin/*.zip /root/out/templates-mono
+    rm -f bin/*.zip
+  fi
 fi
 
 echo "Web build successful"
